requirements.txt
altgraph==0.16.1
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
certifi==2018.1.18
cffi==1.12.3
chardet==3.0.4
coverage==4.5.3
cryptography==2.3
dj-database-url==0.5.0
Django==2.2.4
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.8
django-cors-headers==3.0.2
django-heroku==0.3.1
django-invitations==1.9.2
django-partial-date==1.2.2
django-sendgrid-v5==0.8.0
django-unfriendly==0.6
future==0.17.1
gunicorn==19.9.0
httplib2==0.9.2
idna==2.6
importlib-metadata==0.15
Jinja2==2.10.1
keyring==10.6.0
keyrings.alt==3.0
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
macaroonbakery==1.1.3
macholib==1.11
Mako==1.0.7
MarkupSafe==1.0
more-itertools==7.0.0
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
pbr==5.2.1
pefile==2019.4.18
pexpect==4.2.1
Pillow==5.1.0
pluggy==0.12.0
protobuf==3.0.0
psycopg2==2.8.2
ptyprocess==0.6.0
py==1.8.0
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
pytest==4.5.0
pytest-cov==2.7.1
python-dateutil==2.6.1
python-debian==0.1.32
python-decouple==3.1
python-http-client==3.1.0
pytz==2019.1
pyxdg==0.26
pyyaml==4.2b1
reportlab==3.4.0
requests==2.20.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
sendgrid==6.0.5
simplejson==3.13.2
six==1.12.0
sqlparse==0.3.0
testresources==2.0.1
urllib3==1.24.2
wadllib==1.3.2
wcwidth==0.1.7
whitenoise==4.1.2
wheel==0.33.6
zipp==0.5.1
zope.interface==4.3.2
